+++ data/markdown/report_cancel.md
@@ -1,0 +1,1 @@
+Il report NON è stato inviato
