+++ modules/handlers/__init__.py
@@ -1,22 +1,17 @@
 """Modules that handle the events the bot recognizes and reacts to"""
-from modules.data import config_map
-from .start import start_cmd
+from .ban import ban_cmd
+from .cancel import cancel_cmd
+from .clean_pending import clean_pending_cmd
+from .forwarded_post import forwarded_post_msg
 from .help import help_cmd
+from .job_handlers import clean_pending_job
+from .purge import purge_cmd
+from .reply import reply_cmd
+from .report_spot import report_spot_conv_handler
+from .report_user import report_user_conv_handler
 from .rules import rules_cmd
-
-CHAT_PRIVATE_ERROR = f"Non puoi usare quest comando ora\nChatta con {config_map['bot_tag']} in privato"
-INVALID_MESSAGE_TYPE_ERROR = "Questo tipo di messaggio non è supportato\n"\
-                                "È consentito solo testo, stikers, immagini, audio, video o poll\n"\
-                                "Invia il post che vuoi pubblicare\n"\
-                                "Puoi annullare il processo con /cancel"
-STATE = {
-    'posting': 1,
-    'confirm': 2,
-    'reporting_spot': 3,
-    'reporting_user': 4,
-    'reporting_user_reason': 6,
-    'sending_user_report': 7,
-    'end': -1
-}
-
-purge_flag = False
+from .sban import sban_cmd
+from .settings import settings_cmd
+from .spot import spot_conv_handler
+from .start import start_cmd
+from .stats import stats_callback, stats_cmd
