+++ modules/utils/__init__.py
@@ -1,3 +1,5 @@
 """Modules that provide various util"""
 from .info_util import EventInfo
 from .class_util import Singleton
+from .keyboard_util import get_approve_kb, get_confirm_kb, get_settings_kb, get_stats_kb, get_vote_kb
+from .conversation_util import conv_fail, conv_cancel
